pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-assemblyai-mvp"
version = "0.1.0"
description = "Production-like MVP Telegram bot for audio transcription via AssemblyAI"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot==21.11.1",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "python-docx==1.1.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]